+++ AOS/app/src/main/java/com/example/eeos/presentation/detail/DetailScreen.kt
@@ -19,6 +19,8 @@ import androidx.compose.ui.res.colorResource
 import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.tooling.preview.Preview
 import com.example.eeos.R
+import com.example.eeos.presentation.detail.bottomsheet.BottomSheetContents
+import com.example.eeos.presentation.detail.bottomsheet.SheetDragHandle
 import com.example.eeos.presentation.topappbar.EeosTopAppBar
 
 @OptIn(ExperimentalMaterial3Api::class)
